+++ gridstatus/pjm.py
@@ -101,6 +101,15 @@ class PJM(ISOBase):
 
         Arguments:
             date (datetime.date, str): date to get load for. must be in last 30 days
+
+        Returns:
+            pd.DataFrame: Load data time series. Columns: Time, Load, and all areas
+
+            * Load columns represent PJM-wide load
+            * Returns data for the following areas: AE, AEP, APS, ATSI,
+            BC, COMED, DAYTON, DEOK, DOM, DPL, DUQ, EKPC, JC,
+            ME, PE, PEP, PJM MID ATLANTIC REGION, PJM RTO,
+            PJM SOUTHERN REGION, PJM WESTERN REGION, PL, PN, PS, RECO
         """
 
         if date == "latest":
@@ -114,7 +123,6 @@ class PJM(ISOBase):
             "sort": "datetime_beginning_utc",
             "isActiveMetadata": "true",
             "fields": "area,datetime_beginning_utc,instantaneous_load",
-            "area": "PJM RTO",
         }
         load = self._get_pjm_json(
             "inst_load",
@@ -124,15 +132,24 @@ class PJM(ISOBase):
             verbose=verbose,
         )
 
-        load = load.drop("area", axis=1)
+        # pivot on area
+        load = load.pivot_table(
+            index="Time",
+            columns="area",
+            values="instantaneous_load",
+            aggfunc="first",
+        ).reset_index()
 
-        load = load.rename(
-            columns={
-                "instantaneous_load": "Load",
-            },
-        )
+        load.columns.name = None
 
-        load = load[["Time", "Load"]]
+        # don't need time column
+        all_areas = load.columns.tolist()[1:]
+
+        # set Load column name to match return column of other ISOs
+        load["Load"] = load["PJM RTO"]
+
+        # return everything in correct order
+        load = load[["Time", "Load"] + all_areas]
 
         return load